+++ src/DataObject/IDFilterInput.php
@@ -1,0 +1,39 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+declare(strict_types=1);
+
+namespace OxidEsales\GraphQL\Base\DataObject;
+
+use TheCodingMachine\GraphQLite\Types\ID;
+use Doctrine\DBAL\Query\QueryBuilder;
+
+use function strtoupper;
+
+class IDFilterInput implements FilterInputInterface
+{
+    /** @var ID */
+    private $equals;
+
+    public function __construct(
+        ID $equals
+    ) {
+        $this->equals = $equals;
+    }
+
+    public function equals(): ID
+    {
+        return $this->equals;
+    }
+
+    public function addToQuery(QueryBuilder $builder, string $field): void
+    {
+        $builder->andWhere(strtoupper($field) . ' = :' . $field)
+                ->setParameter(':' . $field, $this->equals);
+        return;
+    }
+}