+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+- `filters` to spec
+- basic input filter types for `string`, `float`, `integer`, `boolean` and `ID`
+- `use function` for all PHP functions used
+
 ## [1.2.0] 2019-12-02
 
 ### Added